Changes to Span exporter with an asynchronous logic:
The write asynchronously send the spans, and recover the failed sent spans in case of error, it alwsys return success
flush sends whatever is left without being sent and waits for the payload to be sent before returning
