+++ Sources/Exporters/OpenTelemetryProtocol/trace/utils/HTTPClient.swift
@@ -28,9 +28,7 @@ internal final class HTTPClient {
         let task = session.dataTask(with: request) { data, response, error in
             completion(httpClientResult(for: (data, response, error)))
         }
-        
         task.resume()
-        Thread.sleep(forTimeInterval: 1)  // FIXME is there a better way to handle this?
     }
 }
